A number of issues related to incorrectly parsing a variety of PLSQL packages. So there are quite a few Oracle specific parsing bugs fixed.
sql.c was very geared for Oracle syntax. This patch has extended the parser to
work with Sybase Adaptive Server Anywhere 9.0.2 and above.  Extending the
syntax means support for more ANSI style SQL.  This could potentially open the
door for DB2 and other syntaxes.

Some of the additional types supported are:
    create view;
    create synonym;
    create index;
    create event;
    create publication;
    create service;
    create domain;
    MobiLink Table Scripts
    MobiLink Conn Scripts

The other item I have added is scope. For example,
with the current sql.c you can turn on "local
variables". But if you have a complex Oracle package
which could easily have 30 different procedures within,
you did not know which variable was defined in which
procedure. Now the local variables are displayed with:
proc_name.local_var_name

This type of scoping was added to a number of objects,
like records and fields. Depends what object was
tagged if adding scope made sense or not.  This was done
to make sifting through all the tags created more manageable.
If your database has hundreds of indicies, I would prefer to 
see them grouped by table name.  Then I can quickly see at a 
glance all the indicies created for a given table.  This of
course assumes you are sorting the output from ctags.

Example for scope of indicies.  
    CREATE INDEX "idx1" ON T1 (
            "id" ASC
    )
Will create an index tag of:
    T1.idx1

Expanded the syntax for existing objects regarding double quotes. 
For example it will recognize:
    CREATE TABLE DBA.T1
    CREATE PROCEDURE "DBA"."P2"
    CREATE VIEW "DBA".V3

So many objects can have owner/schema names (DBA above), the parser now accepts
this. Double quotes should be ignored, but accepted.

When reading records, choose a more appropriate
stopping point since tables can look like this:
    create table t1 (
        c1 integer,
        c2 char(30),
        c3 numeric(10,5),
        c4 integer,
        constraint whatever,
        primary key(c1),
        foreign key (),
        check ()
    )
You do not want to create tags for constraint, primary key and so on.

Support EXCEPTION clauses.
EXCEPTION
<exception handler>;

Where an exception handler could be:
    BEGIN
        WHEN OTHERS THEN
            x := x + 3;
    END;

Prior to this release the old command terminator was a ";".
When creating Oracle packages many different terminators are common:
    ;
    ~
    /
And for SQL Anywhere:
    ;
    go
Now all these command terminators are accepted.  


The following bugs have been fixed by this release:
    1324663 - busy endless spin when parsing .sql file
    1428714 - ctag hangs on open sql comment
    1570779 - problem with SQL parsing such as varchar(20)




git-svn-id: svn://svn.code.sf.net/p/ctags/code/trunk@463 c5d04d22-be80-434c-894e-aa346cc9e8e8